Fix Enums.leq.
The previous implementation that relied on joining both arguments and testing whether the second argument is equal to the join is inherently buggy for Enums.
The reason is that the join does not always return the least upper bound, as computing the least upper bound might be too expensive. Fixes #332