Move solver data serialization logic outside of TD3.
Add the possibility for solver to now additionally return their
state for serialization (as an Obj.t option), that later might be reused
for incremental analysis. Currently TD3 is the only solver making use of this.

The idea is to move the serialization and deserialization logic out of TD3,
so that it later can more easily be adapted, in particular for the following
two purposes:
1. making the serialization no longer dependent on Git-Commit-IDs
2. allowing to keep the data needed for incremental runs in memory, i.e.
so that data does not need to be stored to disk and read from there.