Incremental: Remove reliance on git-commits to version incremental results.
No longer require that the current version of the incrementally analyzed
software is a committed version. Analysis results are no longer stored
in a directory named after the commit, but in some temporary results dir.
After the analysis completed, the temporary results are renamed so
they will be reused in the next incremental analysis run.