Disable hashconsing when incremental.load or incremental.save is active.
The incremental analysis does currently not work for now with the
hashconsing. The one known problem is that the check, whether start states
of queried functions changed does not work with hashconsing, because
different objects (one from the previous and current run) have the same tag.
Thus hashconsing is disabled when incremental.load or incremental.save
is enabled.

The incremental config is changed accordingly.